app/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
lxml==5.1.0
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.31.0
sendgrid==6.11.0
starkbank-ecdsa==2.2.0
urllib3==2.2.0
websockets==12.0
zip-files==0.4.1
